+++ constants/sponsors.ts
@@ -1,0 +1,184 @@
+export const SPONSORSINFO = [
+  {
+    id: 1,
+    name: '鈦坦科技',
+    logo: '/images/sponsors/titansoft.svg',
+    badge: '連續第三年贊助',
+    description:
+      '新加坡商鈦坦科技是一間軟體開發公司,致力於線上軟體平台客製開發與維護,以及代理 Atlassian 旗下產品,以自身實際使用經驗提供一站式導入服務。',
+    socialLinks: [
+      {
+        name: 'web',
+        icon: '/images/icon/web.svg',
+        url: 'https://titansoft.com/tw',
+      },
+      {
+        name: 'fb',
+        icon: '/images/icon/fb.svg',
+        url: 'https://www.facebook.com/TitansoftTW',
+      },
+      {
+        name: 'youtube',
+        icon: '/images/icon/youtube.svg',
+        url: 'https://www.youtube.com/channel/UC_530TqXXizuzA5Ut8wFp4w',
+      },
+      {
+        name: 'ig',
+        icon: '/images/icon/ig.svg',
+        url: 'https://www.instagram.com/titaner_titansoft',
+      },
+    ],
+  },
+  {
+    id: 2,
+    name: '昇新科技',
+    logo: '/images/sponsors/ascentisTech.svg',
+    badge: '',
+    description:
+      'AscentisTech 專注於打造全球領先的線上軟體開發平台,匯集實力堅強、技術與數據背景的菁英團隊,具備高效能系統架構設計、多語系支援與模組化整合能力。',
+    socialLinks: [
+      {
+        name: 'fb',
+        icon: '/images/icon/fb.svg',
+        url: 'https://www.facebook.com/AscentisTech',
+      },
+      {
+        name: 'ig',
+        icon: '/images/icon/ig.svg',
+        url: 'https://www.instagram.com/ascentistech/',
+      },
+    ],
+  },
+]
+
+export const ORGANIZERS = [
+  {
+    id: 1,
+    name: '五倍學院',
+    logo: '/images/sponsors/5xcampus.svg',
+    description: [
+      '五倍學院十餘年來深耕程式教育，陪伴無數學員從零起步，成為引領時代的軟體工程師與 AI 開發人才。',
+      '我們以實戰為核心，打造最貼近產業需求的課程，內容涵蓋前後端網頁技術、AI 工具應用與自動化實作，幫助學員在瞬息萬變的科技浪潮中不斷成長，用技術實現自我價值。',
+      '同時，我們也為企業提供專業的技術顧問與教育訓練服務，協助優化開發流程、強化團隊戰力，讓技術成為驅動組織創新與成長的關鍵力量，提升效率、創造更高價值。',
+    ],
+    socialLinks: [
+      {
+        name: 'web',
+        icon: '/images/icon/web.svg',
+        url: 'https://5xcampus.com/',
+      },
+      {
+        name: 'fb',
+        icon: '/images/icon/fb.svg',
+        url: 'https://www.facebook.com/5xruby',
+      },
+      {
+        name: 'youtube',
+        icon: '/images/icon/youtube.svg',
+        url: 'https://www.youtube.com/c/5xruby-edu',
+      },
+      {
+        name: 'ig',
+        icon: '/images/icon/ig.svg',
+        url: 'https://www.instagram.com/5xcampus/',
+      },
+    ],
+  },
+  {
+    id: 2,
+    name: '六角學院',
+    logo: '/images/sponsors/hexschool.svg',
+    description: [
+      '六角學院是一所專注培養軟體工程師與數位人才的職能學校，目前已累積超過 30,000 名學員。我們的核心使命，是讓更多人透過程式學習，為自己的職涯創造更多選擇與可能性。',
+      '課程橫跨前端、後端到全端開發，強調實戰專題、系統化學習規劃與職涯陪跑機制，不只教會寫程式，更著重溝通力、解決問題能力與自學能力等「工程師思維」，協助學員在 AI 時代持續成長與轉職發展。',
+    ],
+    socialLinks: [
+      {
+        name: 'web',
+        icon: '/images/icon/web.svg',
+        url: 'https://www.hexschool.com/',
+      },
+      {
+        name: 'fb',
+        icon: '/images/icon/fb.svg',
+        url: 'https://www.facebook.com/hexschool',
+      },
+      {
+        name: 'youtube',
+        icon: '/images/icon/youtube.svg',
+        url: 'https://www.youtube.com/@hexschool',
+      },
+      {
+        name: 'ig',
+        icon: '/images/icon/ig.svg',
+        url: 'https://www.instagram.com/hexschool/',
+      },
+    ],
+  },
+  {
+    id: 3,
+    name: '悠識學院',
+    logo: '/images/sponsors/userxper.svg',
+    description: [
+      '悠識學院致力於推動實戰型的 UX（使用者體驗）學習活動，課程主題聚焦使用者訪談、問卷調查、數據分析等專業方法，幫助學員從真實案例中培養洞察與分析能力。我們長期關注設計與科技的交會，定期舉辦以 AI/UX、DesignOps 設計營運與數位職涯發展為主題的講座與實務研討，協助設計師與產品團隊掌握產業趨勢、拓展跨領域視野。',
+      '此外，悠識學院也經營多元社群平台——包含 Circle 社群、IG、Threads，以及 Podcast《UX 有差嗎？》，邀請內部研究員、設計師及同業專家們，分享從事使用者體驗工作的第一線觀察。',
+    ],
+    socialLinks: [
+      {
+        name: 'web',
+        icon: '/images/icon/web.svg',
+        url: 'https://blog.userxper.com/writer/userxper-edu/',
+      },
+      {
+        name: 'fb',
+        icon: '/images/icon/fb.svg',
+        url: 'https://www.facebook.com/userxper.tw',
+      },
+      {
+        name: 'youtube',
+        icon: '/images/icon/youtube.svg',
+        url: 'https://www.youtube.com/@UserXper',
+      },
+      {
+        name: 'ig',
+        icon: '/images/icon/ig.svg',
+        url: 'https://www.instagram.com/userxper/',
+      },
+    ],
+  },
+]
+
+export const CO_ORGANIZERS = [
+  {
+    id: 1,
+    name: '版塊設計',
+    logo: '/images/home/sponsors/BlockStudio.svg',
+    description: [
+      '版塊是網站立體化數位體驗的探險者，從數位策略規劃到品牌視覺識別再到動態特效網站 ，全方位視覺數位形象設計公司。以動態特效為核心、網站設計為起點，為客戶提供全面的數位形象策略解。',
+    ],
+    socialLinks: [
+      {
+        name: 'web',
+        icon: '/images/icon/web.svg',
+        url: 'https://blockstudio.tw',
+      },
+    ],
+  },
+  {
+    id: 2,
+    name: 'WebCRACK',
+    logo: '/images/home/sponsors/WebCRACK.svg',
+    description: [
+      'WebCRACK 是一個專屬於數位設計的線上共享平台，匯聚國內數位設計的能量，創造資源媒合、知識傳遞及作品認證的創意樞紐，構建完整的數位設計生態圈。',
+      '從知識系統建構開始，WebCRACK 聚集設計資源與智慧，為設計師提供實用的設計資訊；透過作品展示認證，讓設計作品被看見並實現價值；更以培育數位人才為目標，透過交流與知識分享，助力未來設計力量的成長。',
+      '我們相信，透過共享交流機制，WebCRACK 不僅促進設計產業升級，更提升整體產業品質，為數位設計注入無限可能。',
+    ],
+    socialLinks: [
+      {
+        name: 'web',
+        icon: '/images/icon/web.svg',
+        url: 'https://webcrack.tw/',
+      },
+    ],
+  },
+]
